better file structure for page transitions
